moved curve
